+++ src/widgets/busy-button/busy-button.tsx
@@ -3,13 +3,16 @@ import React from "react";
 import { BusyButton as IBusyButton } from "../../../types";
 import { Loader } from "../loader/loader";
 
-export const BusyButton = ({ busy, disabled, className, children, ...rest }: IBusyButton): any => {
+export const BusyButton = ({ busy, invert, disabled, className: _className, children, ...rest }: IBusyButton): any => {
+    const className = (): string => {
+        const classes = ["react-simple-widget", "busy-button"];
+        if (_className) classes.push(_className);
+        return classes.join(" ");
+    };
+
     return (
-        <button
-            className={`react-simple-widget busy-button ${className ? className : ""}`}
-            disabled={busy || disabled}
-            {...rest}>
-            {busy && <Loader className="d-inline" role="status" aria-hidden="true" />}
+        <button className={className()} disabled={busy || disabled} {...rest}>
+            {busy && <Loader className="d-inline" role="status" aria-hidden="true" invert={invert} />}
             {children}
         </button>
     );