+++ src/widgets/confirm-button/confirm-button.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { ConfirmButton as IConfirmButton } from "../../../types";
 import { DialogProviderContext } from "../dialog-provider/dialog-provider-context";
-import { ConfirmDialog } from "./confirm-dialog";
+import { ConfirmButtonDialog } from "./confirm-button-dialog";
 import { BusyButton } from "../busy-button/busy-button";
 
 export const ConfirmButton = ({
@@ -17,13 +17,16 @@ export const ConfirmButton = ({
     const { showDialog } = useContext(DialogProviderContext);
 
     const confirm = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>): void => {
-        showDialog(helper => <ConfirmDialog helper={helper} message={message} confirmButtonClassName={className} />, {
-            backgroundDismissible: true,
-            onDismissed: (proceed: boolean): void => {
-                if (proceed) return onConfirm();
-                if (onCancel) return onCancel();
+        showDialog(
+            helper => <ConfirmButtonDialog helper={helper} message={message} confirmButtonClassName={className} />,
+            {
+                backgroundDismissible: true,
+                onDismissed: (proceed: boolean): void => {
+                    if (proceed) return onConfirm();
+                    if (onCancel) return onCancel();
+                }
             }
-        });
+        );
 
         if (_onClick) _onClick(e);
     };