+++ src/widgets/month-date-picker/month-date-picker.stories.tsx
@@ -7,10 +7,17 @@ djs.extend(advancedFormat);
 
 export default { title: "MonthDatePicker", component: MonthDatePicker };
 
-export const usage = (): any => {
+export const Default = (): any => {
     const ExampleApp = (): any => {
         const [date, setDate] = useState("2021-09-04");
-        return <MonthDatePicker className="d-inline-block" onDateSelect={setDate} date={date} />;
+
+        return (
+            <MonthDatePicker
+                className="btn btn-outline-primary btn-sm d-inline-block w-auto"
+                onDateSelect={setDate}
+                date={date}
+            />
+        );
     };
 
     return <ExampleApp />;